Rename StoryPostHero to StoryHero and change useStoryHero condition
